[CI] add features to LLVM
Unfortunately we don't get any **prebuild packages** for `IWYU` from
the _official site_ or _winget_ [ or _chocolatey_ ] for windows. We will
have to compile `IWYU` from the source to run the scans.

`clang` and `clang-tools-extra` features should be compiled along with
`LLVM` to compile `IWYU`. Doing so makes resolving issue **1183** of `WasmEdge` easier.

Signed-off-by: abhinandanudupa <67597710+abhinandanudupa@users.noreply.github.com>